refactor: 패키지 이동 player와 관련된 도메인만 패키지로 모았다.
